BBC Front ender is op
No I'm human is the game where the art draws in you in. The lore is great. And the atmosphere is amazing. Too bad the gameplay is trash and its variety is not different enough to even get the other endings. It just a game where you watch on youtube. short 4-5 hour game for 15 dollars. And the gameplay is just is this guy a vistor? Shot him or not. That it